Implement custom lexer

+++ src/ast.rs
@@ -92,7 +92,7 @@ pub enum Constant {
 
 #[derive(Debug, Clone)]
 pub enum NonString {
-    Integer(u32),
+    Integer(usize),
     Real(f32),
     Identifier(Identifier),
 }
@@ -106,7 +106,7 @@ pub enum UnsignedConstant {
 
 #[derive(Debug)]
 pub enum Number {
-    Integer(u32),
+    Integer(usize),
     Real(f32),
 }
